# server/openapi.yaml
openapi: 3.0.3
info:
  title: National Dialogue ZA - Admin Portal API
  description: System of record API for South Africa's National Dialogue programme
  version: 1.0.0
  contact:
    name: National Dialogue ZA
    email: admin@nationaldialogue.za
servers:
  - url: http://localhost:5000/api
    description: Development server
  - url: https://admin.nationaldialogue.za/api
    description: Production server
paths:
  /auth/login:
    post:
      summary: User authentication
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthLoginRequest'
      responses:
        '200':
          description: Successful authentication
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthLoginResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/me:
    get:
      summary: Get current user information
      tags: [Authentication]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users:
    get:
      summary: List users with optional filtering
      tags: [Users]
      security:
        - bearerAuth: []
      parameters:
        - name: role
          in: query
          schema:
            type: string
          description: Filter by role
        - name: active
          in: query
          schema:
            type: boolean
          description: Filter by active status
        - name: search
          in: query
          schema:
            type: string
          description: Search by name or email
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      summary: Create a new user
      tags: [Users]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/{id}:
    get:
      summary: Get user by ID
      tags: [Users]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    patch:
      summary: Update user
      tags: [Users]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: Delete user
      tags: [Users]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: User deleted successfully
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /submissions:
    get:
      summary: List submissions with filtering and pagination
      tags: [Submissions]
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/SubmissionStatus'
        - name: province
          in: query
          schema:
            $ref: '#/components/schemas/Province'
        - name: channel
          in: query
          schema:
            $ref: '#/components/schemas/Channel'
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of submissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmissionsResponse'

    post:
      summary: Create a new submission
      tags: [Submissions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSubmissionRequest'
      responses:
        '201':
          description: Submission created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Submission'
  /submissions/{id}:
    get:
      summary: Get submission by ID
      tags: [Submissions]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Submission details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Submission'

    patch:
      summary: Update submission (moderation actions)
      tags: [Submissions]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSubmissionRequest'
      responses:
        '200':
          description: Submission updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Submission'
  /cases:
    get:
      summary: List cases with filtering and pagination
      tags: [Cases]
      security:
        - bearerAuth: []
      parameters:
        - name: state
          in: query
          schema:
            $ref: '#/components/schemas/CaseState'
        - name: department
          in: query
          schema:
            type: string
            format: uuid
        - name: assignee
          in: query
          schema:
            type: string
            format: uuid
        - name: priority
          in: query
          schema:
            $ref: '#/components/schemas/Priority'
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of cases
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CasesResponse'

    post:
      summary: Create a new case
      tags: [Cases]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCaseRequest'
      responses:
        '201':
          description: Case created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Case'
  /cases/{id}:
    get:
      summary: Get case by ID
      tags: [Cases]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Case details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Case'

    patch:
      summary: Update case
      tags: [Cases]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCaseRequest'
      responses:
        '200':
          description: Case updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Case'
  /departments:
    get:
      summary: List departments
      tags: [Departments]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of departments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Department'

    post:
      summary: Create a new department
      tags: [Departments]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDepartmentRequest'
      responses:
        '201':
          description: Department created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Department'
  /departments/{id}:
    get:
      summary: Get department by ID
      tags: [Departments]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Department details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Department'

    patch:
      summary: Update department
      tags: [Departments]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDepartmentRequest'
      responses:
        '200':
          description: Department updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Department'
  /analytics/summary:
    get:
      summary: Get analytics summary
      tags: [Analytics]
      security:
        - bearerAuth: []
      parameters:
        - name: range
          in: query
          schema:
            type: string
            enum: [7d, 30d, 90d]
            default: 7d
      responses:
        '200':
          description: Analytics summary data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsSummary'
  /polls:
    get:
      summary: List polls
      tags: [Polls]
      security:
        - bearerAuth: []
      parameters:
        - name: active
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: List of polls
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Poll'

    post:
      summary: Create a new poll
      tags: [Polls]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePollRequest'
      responses:
        '201':
          description: Poll created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Poll'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AuthLoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: "thandi@admin.local"
        password:
          type: string
          example: "admin123"
    AuthLoginResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/UserInfo'
        accessToken:
          type: string
        refreshToken:
          type: string
...
    User:
      type: object
      required: [id, name, email, isActive, createdAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: "Thandi Dlamini"
        email:
          type: string
          format: email
          example: "thandi@admin.local"
        phone:
          type: string
          nullable: true
          example: "+27123456789"
        province:
          $ref: '#/components/schemas/Province'
        isActive:
          type: boolean
          example: true
        roles:
          type: array
          items:
            type: string
          example: ["SuperAdmin"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateUserRequest:
      type: object
      required: [name, email, passwordHash, roles]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        province:
          $ref: '#/components/schemas/Province'
        passwordHash:
          type: string
        isActive:
          type: boolean
          default: true
        roles:
          type: array
          items:
            type: string
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        province:
          $ref: '#/components/schemas/Province'
        isActive:
          type: boolean
        roles:
          type: array
          items:
            type: string

    # Submission Schemas
...
    CreateSubmissionRequest:
      type: object
      required: [text, channel]
      properties:
        userId:
          type: string
          format: uuid
        channel:
          $ref: '#/components/schemas/Channel'
        text:
          type: string
        mediaUrls:
          type: array
          items:
            type: string
            format: uri
        languageDetected:
          type: string
        province:
          $ref: '#/components/schemas/Province'
    UpdateSubmissionRequest:
      type: object
      properties:
        status:
          $ref: '#/components/schemas/SubmissionStatus'
        sentiment:
          type: number
          minimum: -1
          maximum: 1
        toxicity:
          type: boolean
    SubmissionsResponse:
      type: object
      properties:
        submissions:
          type: array
          items:
            $ref: '#/components/schemas/Submission'
        total:
          type: integer

    # Case Schemas
...
    SubmissionStatus:
      type: string
      enum: [new, moderated, routed, in_progress, resolved, declined]
    Channel:
      type: string
      enum: [mobile, web, whatsapp, social]
    CaseState:
      type: string
      enum: [open, investigating, awaiting_info, resolved]
    Priority:
      type: string
      enum: [low, medium, high]
...
    Province:
      type: string
      enum: [eastern_cape, free_state, gauteng, kwazulu_natal, limpopo, mpumalanga, northern_cape, north_west, western_cape]

    # Error Schema
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: "VALIDATION_ERROR"
            message:
              type: string
              example: "Invalid request data"
            details:
              type: object
              description: "Additional error details"
...